Break the build by declaring bar's public API
This step *should* break the build, but does not, because while the the
API specification DSL is implemented, fully handling it by generating
api and impl jars is not. When that work is complete, this step will
fail for the *correct* reason that bar's public API will be being
enforced.